+++ data/task-list.txt
@@ -1,6 +1,6 @@
 1.[E][X] conference (at:  thursday 4pm)
-2.[T][X] read
-3.[D][ ] dive (by: 2021-12-31 14:00)
-4.[T][ ] swim
-5.[E][ ] concert (at:  monday 12-2am)
-6.[D][ ] renew (by: 2021-09-23 18:00)
+2.[D][ ] dive (by: 2021-12-31 14:00)
+3.[T][ ] swim
+4.[E][X] concert (at:  monday 12-2am)
+5.[D][ ] renew (by: 2021-09-23 18:00)
+6.[D][X] register (by: 2021-10-12 16:00)
